+++ packages/components/CHANGELOG.md
@@ -9,15 +9,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `Breakpoint` component
 - `Tree / TreeItem`
   - disabled and selected states
   - `brand` prop
   - `treeBackgroundColor` util function
 - `ProgressCircular` component
+- `UnorderedList`, `OrderedList`
+
 
 ### Changed
 
 - `Tree` now uses the same `selected` color as `TreeItem`
+- `DialogHeader` uses responsive padding values
+- `CardMedia` supports background color props
+- `InputDate` & `InputDateRange` vertial margin removed to be consistent with other inputs
 
 ### Fixed